# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ExamForge API
  version: 1.0.0
  description: |
    Tài liệu Swagger/OpenAPI cho toàn bộ API đã implement trong các module.

    Ghi chú:
    - API gốc được mount dưới `/api/v1`.
    - Các endpoint thuộc nhóm tenant-based sẽ ưu tiên resolve tenant qua subdomain, sau đó mới đọc header `X-Tenant-ID`.
    - Các lỗi `422` là lỗi validation chuẩn của Laravel.
    - Một số lỗi nghiệp vụ ở `QuestionService` đang được ném bằng `\Exception`; nếu chưa map ở global exception handler thì response thực tế có thể cần chuẩn hoá thêm.
servers:
  - url: /api/v1
    description: API v1
...
security:
  - BearerAuth: []
paths:
...
  /questions/:
    get:
      tags: [Questions]
      summary: Danh sách câu hỏi
      description: Lấy danh sách câu hỏi theo tenant hiện tại, hỗ trợ lọc theo type, status, tags và từ khóa tìm kiếm.
      parameters:
        - $ref: '#/components/parameters/XTenantID'
        - $ref: '#/components/parameters/QuestionTypeQuery'
        - $ref: '#/components/parameters/QuestionStatusQuery'
        - $ref: '#/components/parameters/TagsQuery'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Danh sách câu hỏi
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/QuestionListData'
        '400':
          $ref: '#/components/responses/BadRequestApiError'
        '401':
          $ref: '#/components/responses/UnauthorizedMessage'
        '404':
          $ref: '#/components/responses/NotFoundError'
    post:
      tags: [Questions]
      summary: Tạo câu hỏi
      description: Tạo câu hỏi mới, mặc định trạng thái là `draft`.
      parameters:
        - $ref: '#/components/parameters/XTenantID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionCreateRequest'
      responses:
        '201':
          description: Tạo câu hỏi thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/QuestionResource'
        '400':
          $ref: '#/components/responses/BadRequestApiError'
        '401':
          $ref: '#/components/responses/UnauthorizedMessage'
        '403':
          $ref: '#/components/responses/ForbiddenMessage'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  parameters:
    XTenantID:
      name: X-Tenant-ID
      in: header
      description: |
        UUID của tenant hiện tại.
        Middleware sẽ ưu tiên resolve theo subdomain, sau đó mới đọc header này.
      required: false
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      description: Số trang.
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPage:
      name: per_page
      in: query
      description: Số bản ghi mỗi trang.
      schema:
        type: integer
        minimum: 1
        default: 15
...
    SearchQuery:
      name: search
      in: query
      description: Từ khóa tìm theo email hoặc tên hiển thị.
      schema:
        type: string
        maxLength: 255
    QuestionTypeQuery:
      name: type
      in: query
      description: Lọc câu hỏi theo loại câu hỏi.
      schema:
        type: string
        enum: [multiple_choice, short_answer, essay, true_false]
    QuestionStatusQuery:
      name: status
      in: query
      description: Lọc câu hỏi theo trạng thái.
      schema:
        type: string
        enum: [draft, published, archived]
    TagsQuery:
      name: tags
      in: query
      description: Danh sách tag. Có thể truyền lặp nhiều lần, ví dụ `tags=math&tags=algebra`.
      required: true
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
...
  responses:
    UnauthorizedMessage:
      description: Chưa xác thực
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageOnlyResponse'
          example:
            message: Unauthenticated
    ForbiddenMessage:
      description: Không có quyền truy cập
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageOnlyResponse'
          example:
            message: Forbidden
...
    NotFoundError:
      description: Không tìm thấy dữ liệu
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: User not found
            errors: null
    BadRequestApiError:
      description: Lỗi nghiệp vụ hoặc dữ liệu đầu vào không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Tenant not resolved
            errors: null
    ValidationError:
      description: Dữ liệu gửi lên không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            message: The given data was invalid.
            errors:
              email:
                - Email is already registered.
  schemas:
    SuccessEnvelope:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Thành công
        data:
          nullable: true
    ErrorEnvelope:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Có lỗi xảy ra
        errors:
          nullable: true
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items: {}
    MessageOnlyResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Unauthorized
    ValidationErrorResponse:
      type: object
      required:
        - message
        - errors
      properties:
        message:
          type: string
          example: The given data was invalid.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    QuestionResource:
      type: object
      required:
        - id
        - tenant_id
        - created_by
        - type
        - content
        - options
        - correct_answer
        - max_score
        - difficulty
        - tags
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        type:
          type: string
          enum: [multiple_choice, short_answer, essay, true_false]
        content:
          type: string
        options:
          type: array
          items:
            type: string
          nullable: true
        correct_answer:
          type: array
          items: {}
          nullable: true
        max_score:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, published, archived]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    QuestionListData:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/QuestionResource'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      required: [total, count, per_page, current_page, last_page]
      properties:
        total:
          type: integer
        count:
          type: integer
        per_page:
          type: integer
        current_page:
          type: integer
        last_page:
          type: integer
        from:
          type: integer
          nullable: true
        to:
          type: integer
          nullable: true
...
    QuestionCreateRequest:
      type: object
      required: [type, content, max_score, difficulty]
      properties:
        type:
          type: string
          enum: [multiple_choice, short_answer, essay, true_false]
        content:
          type: string
        options:
          type: array
          nullable: true
          items:
            type: string
        correct_answer:
          type: array
          nullable: true
          items: {}
        max_score:
          type: integer
          minimum: 1
        difficulty:
          type: string
          enum: [easy, medium, hard]
        tags:
          type: array
          nullable: true
          items:
            type: string
...